Move Android PF setup to own file
