+++ unit-tests/src/org/commcare/CommCareTestApplication.java
@@ -6,10 +6,10 @@ import android.util.Log;
 import org.commcare.android.util.TestUtils;
 import org.commcare.dalvik.BuildConfig;
 import org.commcare.models.AndroidPrototypeFactory;
-import org.commcare.models.database.DbUtil;
 import org.commcare.models.database.HybridFileBackedSqlStorage;
 import org.commcare.models.database.HybridFileBackedSqlStorageMock;
 import org.commcare.android.database.app.models.UserKeyRecord;
+import org.commcare.models.database.AndroidPrototypeFactorySetup;
 import org.commcare.services.CommCareSessionService;
 import org.javarosa.core.model.User;
 import org.javarosa.core.services.storage.Persistable;
@@ -98,7 +98,7 @@ public class CommCareTestApplication extends CommCareApplication {
                         .replace("/", ".")
                         .replace(".class", "")
                         .replace(".class", "");
-                DbUtil.loadClass(className, externClasses);
+                AndroidPrototypeFactorySetup.loadClass(className, externClasses);
             }
         } catch (Exception e) {
             Log.w(TAG, e.getMessage());
